Move larpake renderer to separate file
